pyproject.toml
[project]
name = "function-call"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cobra>=0.30.0",
    "openai-agents>=0.6.3",
    "python-dotenv>=1.0.0",
    "streamlit>=1.40.0",
]
